Move date filter into transform module
Rather than using the flawed eachSchema function, this uses the
transform module to filter the document on a per-request basis.
