+++ src/mongoose/json-transform.js
@@ -18,34 +18,6 @@ function jsonTransformPlugin(schema) {
 function filterFields(doc, ret, options) {
   ret = filter(doc, ret, options);
   ret = translateDoc(doc, ret, options);
-  ret = filterDates(doc, ret, options);
-  return ret;
-}
-
-function filterDates(doc, ret, options) {
-  if (!options.at) {
-    return ret;
-  }
-
-  function checkDates(field) {
-    if (!field.start_date && !field.end_date) {
-      return true;
-    }
-    var start = new Date(field.start_date);
-    var end = new Date(field.end_date);
-    var at = options.at;
-
-    return start < at && (!field.end_date || end > at);
-  }
-
-  if (doc.other_names) {
-    ret.other_names = doc.other_names.filter(checkDates);
-  }
-
-  if (doc.memberships) {
-    ret.memberships = doc.memberships.filter(checkDates);
-  }
-
   return ret;
 }
